Correcting bugs in boxing
Now boxing works for small 3d systems.
2d still needs some fixing.

+++ LF_DEM/Box.h
@@ -91,7 +91,7 @@ class Box{
   }
 
 
-  int container_size(){
+  size_t container_size(){
 	return container->size();
   }
